fix(precedence): Use sub instead of method for class method get_table
PrecedenceTable::get_table() is called as a class method
(ClassName->get_table()), but was defined with `method` keyword which
requires an instance. Changed to `sub` to allow class method invocation.

This fixes CI failures in t/cli/app-options.t and t/debug-precedence.t
which showed: "Cannot invoke method 'get_table' on a non-instance"